Electrical database implemented now faster

+++ app/src/main/java/com/example/mobile1_tp3/electricalTerminals/ElectricalTerminal.java
@@ -4,10 +4,10 @@ public class ElectricalTerminal
 {
     private Long id;
     private String name;
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
 
-    public ElectricalTerminal(String nomBorne, Float latitude, Float longitude)
+    public ElectricalTerminal(String nomBorne, Double latitude, Double longitude)
     {
         this.name = nomBorne;
         this.latitude = latitude;
@@ -29,22 +29,22 @@ public class ElectricalTerminal
 
     public void setName(String nomBorne) { this.name = nomBorne; }
 
-    public Float getLatitude()
+    public Double getLatitude()
     {
         return latitude;
     }
 
-    public void setLatitude(Float latitude)
+    public void setLatitude(Double latitude)
     {
         this.latitude = latitude;
     }
 
-    public Float getLongitude()
+    public Double getLongitude()
     {
         return longitude;
     }
 
-    public void setLongitude(Float longitude)
+    public void setLongitude(Double longitude)
     {
         this.longitude = longitude;
     }
